Refactor seperate calls into api namespace
